stop/canStop: Return errors from go API
When used on older macOS version, canStop() will return false, which is
correct, so I did not add an additional error to indicade that this
method is not supported on older macOS releases.

Stop() receives a completion handler which has an err parameter, so I
chose to report the error through it rather than adding a second error
path to the Stop() call.